% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIO.R
\name{get_inndata}
\alias{get_inndata}
\title{Download and store inndata from the NOFA database.}
\usage{
get_inndata(serveradress, datafolder)
}
\arguments{
\item{serveradress}{URL of the server wher the database is located.}

\item{datafolder}{Relative or absolute path of the folder where the data should be stored locally.}
}
\value{
A data.frame stored as .rds object named view_occurrence_by_event.rds
}
\description{
Download and store inndata from the NOFA database.
}
\examples{
get_inndata(serveradress="my-server.no",datafolder="./data")

}
