% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_and_slice}
\alias{wrangle_and_slice}
\title{Wrangle input data for format required by simulation}
\usage{
wrangle_and_slice(start_year, end_year, inndata, focal_species,
  geoselect_native)
}
\arguments{
\item{start_year}{Year of staring recording as introductions}

\item{end_year}{Year of ending recording as introductions}

\item{inndata}{data-frame with inndata comming from "get_inndata"}

\item{focal_species}{The latin (canonical) name of the species}

\item{geoselect_native}{A spatial.polygon.dataframe indicating the native distribution range of the species in epsg:4326}
}
\value{
A data.frame
}
\description{
Wrangle input data for format required by simulation
}
\examples{

setwd("~/temp") # simulation working directory
# Download example input files
get_inndata(serveradress="my-server-adress.no",datafolder="./data") # will take som time to load
download.file("https://ntnu.box.com/shared/static/kzgmsa898gcks57iacogpet1g2yue8ib.rds",
             "./data/geoselect_native_Esox_lucius.rds") # not yet procedures to draw this rigth from DB

inndata <- readRDS("./data/view_occurrence_by_event.rds")
geoselect_native <- readRDS("./data/geoselect_native_Esox_lucius.rds")
focal_species <- "Esox lucius"
mynewdata_aslist <- wrangle_and_slice(inndata=inndata,start_year=1970,end_year=2017)
mynewdata <- mynewdata_aslist[["data"]]

}
